Handle plot axes if only 1 dataSrc
This is what I came up with as a quick hack, after coming back to
this program after almost a year later. Because EU dataset is
no longer available from the url, and I had to look at Indian
data (which is still available from the url). And the plot logic
had not been setup to account for only one of these data sources
being available.

NOTE: THe root cause is matplotlib.pyplot returns a 1D array or
2D array based on whether one is requesting only a 1xN or Nx1 matrix
or MxN matrix of plots through its subplot.

However after I got the base data analysed and plotted, thought a
bit and realised the better way of doing it, which I will be
doing next.

This captured here, just as a note. I will be reverting this next.